move utils package to internal/
