+++ internal/staticanalysis/parsing/string_regexp.go
@@ -1,7 +1,7 @@
 package parsing
 
 import (
-	"github.com/ossf/package-analysis/internal/staticanalysis/obfuscation/utils"
+	"github.com/ossf/package-analysis/internal/utils"
 	"os"
 	"regexp"
 	"strings"
